Moved off the shelf mutation functions to mutation.py
